Fixed the nvme storage

+++ main/tensor/builtin_formats.cc
@@ -18,7 +18,7 @@ tensor_creation_fs_t bbts::dense_tensor_t::get_creation_fs() {
   // return the size
   auto size = [](const tensor_meta_t &_meta) {
     auto &m = *(dense_tensor_meta_t *) &_meta;
-    return sizeof(tensor_t) + m.m().num_cols * m.m().num_rows * sizeof(float);
+    return sizeof(tensor_t) + (size_t) m.m().num_cols * (size_t) m.m().num_rows * sizeof(float);
   };
 
   auto pnt = [](const void *here, std::stringstream &ss) {
